localStorageに関する処理をまとめたクラスを追加
