+++ show.html
@@ -11,7 +11,7 @@
     <a href="top.html">投稿する</a>
     <ul id="post-list"></ul>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script src="post-show.js"></script>
-    <script src="post-delete.js"></script>
+    <script type="module" src="post-show.js"></script>
+    <script type="module" src="post-delete.js"></script>
   </body>
 </html>
