+++ top.html
@@ -18,6 +18,6 @@
       <input type="submit" id="post-submit" value="投稿">
     </form>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-    <script src="post-form.js"></script>
+    <script type="module" src="post-form.js"></script>
   </body>
 </html>
